moved crosstalk support classes to utils_crosstalk
