+++ src/qililab/qprogram/__init__.pyi
@@ -21,14 +21,17 @@ if TYPE_CHECKING:
     from .qblox_compiler import QbloxCompilationOutput, QbloxCompiler
     from .qdac_compiler import QdacCompilationOutput, QdacCompiler
     from .qprogram import QProgram, QProgramCompilationOutput
+    from .utils_crosstalk import CrosstalkElements, NonLinearState
 
 __all__ = [
     "Calibration",
+    "CrosstalkElements",
     "CrosstalkMatrix",
     "Experiment",
     "FluxVector",
     "NonLinearCrosstalkMatrix",
     "NonLinearFluxVector",
+    "NonLinearState",
     "QProgram",
     "QProgramCompilationOutput",
     "QbloxCompilationOutput",
